Factorize duplicated Node components
