+++ packages/twenty-front/src/modules/workflow/workflow-diagram/workflow-nodes/components/WorkflowDiagramStepNodeEditable.tsx
@@ -47,7 +47,6 @@ export const WorkflowDiagramStepNodeEditable = ({
     <WorkflowDiagramStepNodeEditableContent
       id={id}
       data={data}
-      variant="default"
       selected={selected ?? false}
       onClick={() => {
         if (!isInRightDrawer) {
