+++ packages/twenty-front/src/modules/workflow/workflow-diagram/workflow-nodes/components/WorkflowDiagramStepNodeEditableContent.tsx
@@ -16,7 +16,6 @@ import { capitalize } from 'twenty-shared/utils';
 import { IconTrash } from 'twenty-ui/display';
 import { FloatingIconButton } from 'twenty-ui/input';
 import { useConnectionState } from '@/workflow/workflow-diagram/workflow-nodes/hooks/useConnectionState';
-import { css } from '@emotion/react';
 import { WorkflowDiagramHandleTarget } from '@/workflow/workflow-diagram/workflow-nodes/components/WorkflowDiagramHandleTarget';
 import { WorkflowDiagramHandleSource } from '@/workflow/workflow-diagram/workflow-nodes/components/WorkflowDiagramHandleSource';
 
@@ -32,44 +31,6 @@ const StyledAddStepButtonContainer = styled.div<{
   left: 50%;
   bottom: 0;
   transform: translateX(-50%) translateY(100%);
-`;
-
-const StyledNodeContainer = styled(WorkflowNodeContainer)<{
-  isConnectable?: boolean;
-}>`
-  border-color: ${({ theme }) => theme.border.color.strong};
-  background: ${({ theme }) => theme.background.secondary};
-
-  &:hover {
-    background: linear-gradient(
-        0deg,
-        ${({ theme }) => theme.background.transparent.lighter} 0%,
-        ${({ theme }) => theme.background.transparent.lighter} 100%
-      ),
-      ${({ theme }) => theme.background.secondary};
-    ${({ theme, isConnectable }) =>
-      isConnectable &&
-      css`
-        border-color: ${theme.color.blue} !important;
-      `};
-  }
-
-  .selected & {
-    border-color: ${({ theme }) => theme.color.blue};
-    background: ${({ theme }) => theme.adaptiveColors.blue1};
-  }
-`;
-
-const StyledNodeLabel = styled(WorkflowNodeLabel)`
-  color: ${({ theme }) => theme.font.color.tertiary};
-
-  .selected & {
-    color: ${({ theme }) => theme.tag.text.blue};
-  }
-`;
-
-const StyledNodeTitle = styled(WorkflowNodeTitle)`
-  color: ${({ theme }) => theme.font.color.primary};
 `;
 
 const StyledDeleteButtonContainer = styled.div`
@@ -114,7 +75,7 @@ export const WorkflowDiagramStepNodeEditableContent = ({
 
   return (
     <>
-      <StyledNodeContainer
+      <WorkflowNodeContainer
         data-click-outside-id={WORKFLOW_DIAGRAM_STEP_NODE_BASE_CLICK_OUTSIDE_ID}
         onClick={onClick}
         onMouseEnter={handleMouseEnter}
@@ -128,10 +89,10 @@ export const WorkflowDiagramStepNodeEditableContent = ({
 
         <WorkflowNodeRightPart>
           <WorkflowNodeLabelWithCounterPart>
-            <StyledNodeLabel>{capitalize(data.nodeType)}</StyledNodeLabel>
+            <WorkflowNodeLabel>{capitalize(data.nodeType)}</WorkflowNodeLabel>
           </WorkflowNodeLabelWithCounterPart>
 
-          <StyledNodeTitle>{data.name}</StyledNodeTitle>
+          <WorkflowNodeTitle highlight>{data.name}</WorkflowNodeTitle>
         </WorkflowNodeRightPart>
 
         {selected && (
@@ -143,7 +104,7 @@ export const WorkflowDiagramStepNodeEditableContent = ({
             />
           </StyledDeleteButtonContainer>
         )}
-      </StyledNodeContainer>
+      </WorkflowNodeContainer>
 
       {!data.hasNextStepIds && !isInProgressConnection && (
         <StyledAddStepButtonContainer
