+++ packages/twenty-front/src/modules/workflow/workflow-diagram/workflow-nodes/components/WorkflowNodeContainer.tsx
@@ -1,6 +1,12 @@
 import styled from '@emotion/styled';
+import type { WorkflowRunStepStatus } from '@/workflow/types/Workflow';
+import { css } from '@emotion/react';
+import { getWorkflowDiagramColors } from '@/workflow/workflow-diagram/utils/getWorkflowDiagramColors';
 
-const StyledNodeContainer = styled.div`
+const StyledNodeContainer = styled.div<{
+  runStatus?: WorkflowRunStepStatus;
+  isConnectable?: boolean;
+}>`
   align-items: center;
   display: flex;
   gap: ${({ theme }) => theme.spacing(2)};
@@ -14,6 +20,34 @@ const StyledNodeContainer = styled.div`
   cursor: pointer;
   position: relative;
   transition: border-color 0.1s;
+
+  &:hover {
+    background: linear-gradient(
+        0deg,
+        ${({ theme }) => theme.background.transparent.lighter} 0%,
+        ${({ theme }) => theme.background.transparent.lighter} 100%
+      ),
+      ${({ theme }) => theme.background.secondary};
+    ${({ theme, isConnectable }) =>
+      isConnectable &&
+      css`
+        border-color: ${theme.color.blue} !important;
+      `};
+  }
+
+  ${({ theme, runStatus }) => {
+    const colors = getWorkflowDiagramColors({ theme, runStatus });
+
+    return css`
+      border-color: ${colors.unselected.borderColor};
+      background: ${colors.unselected.background};
+
+      .selected & {
+        background-color: ${colors.selected.background};
+        border: 1px solid ${colors.selected.borderColor};
+      }
+    `;
+  }}
 `;
 
 export { StyledNodeContainer as WorkflowNodeContainer };
